fix(rpc): Make RPC "incorrect parameters" error code match `zcashd` (#6066) - manually delete getblocktemplate-rpcs
* Move RPC method constants into their own module

* Rename RPC compatibility modules to avoid confusion

* Rename RPC middleware to include its new functionality

* Use FutureExt::inspect() for logging, and only format on failure

* Log all RPC errors at info level

* Make "invalid parameters" RPC error code match `zcashd`
